Added screen messages for each action that can be done.

+++ Source/Node/diskNode.cpp
@@ -249,10 +249,11 @@ void DiskNode::onReadyRead() {
 
                 if (messageFormat.getAction() == ActionMessage::Upload) {
 
-                    if (!fileNamesAdded.contains(messageFormat.getFileName()))
+                    QString nombre = fileName.split("_").at(0); 
+                    if (!fileNamesAdded.contains(nombre))
                     {
-                        fileNamesAdded.append(messageFormat.getFileName());
-
+                        fileNamesAdded.append(nombre);
+                        qDebug() << "Se adiciona:" + nombre + ".";
                     }
 
                     bool success = storeFile(messageFormat.getContent(), fileName);
@@ -263,20 +264,38 @@ void DiskNode::onReadyRead() {
                 }
 
                 else if (messageFormat.getAction() == ActionMessage::Erase) {
+                    if (fileNamesAdded.contains(fileName))
+                    {
+                        deleteFile(fileName); 
 
-                    deleteFile(fileName); 
-
-                    data = "Se borra el pdf: " + fileName.toUtf8();
-                    sendData(buildMessage(MessageIndicator::NodeToController, fileName, ActionMessage::Erase, data));
+                        data = "Se borra el pdf: " + fileName.toUtf8();
+                        sendData(buildMessage(MessageIndicator::NodeToController, fileName, ActionMessage::Erase, data));
+                    }
+                    else
+                    {
+                        data = "No se encontro el siguiente PDF para borrar: " + fileName.toUtf8();
+                        sendData(buildMessage(MessageIndicator::NodeToController, fileName, ActionMessage::Erase, data)); 
+                    }
                 }
 
                 else if (messageFormat.getAction() == ActionMessage::Download) {
-                    qDebug() << "El nodo empieza a mandar info para descarga de pdf";
-                    searchAndSendPdfBlocks(path,messageFormat.getFileName());
+                    if (fileNamesAdded.contains(fileName))
+                    {
+                        qDebug() << "El nodo empieza a mandar info para descarga de pdf";
+                        searchAndSendPdfBlocks(path,messageFormat.getFileName());
+                    }
+                    else
+                    {
+                        QString message = "El pdf " + fileName + " no ha sido subido para poder descargarlo.";
+                        sendData(buildMessage(MessageIndicator::NodeToController, fileName, ActionMessage::Error, message.toUtf8()));
+                
+                    }
+
                 }
 
                 else if (messageFormat.getAction() == ActionMessage::Check) {
                     qDebug() << "Se verifica estatus de pdf: " + fileName;
+
 
                     if (!fileNamesAdded.contains(fileName))
                     {
@@ -288,7 +307,6 @@ void DiskNode::onReadyRead() {
                         data = "El pdf está disponible";
                     }
 
-                    qDebug() << "Pdf disponible";
                     sendData(buildMessage(MessageIndicator::NodeToController, fileName, ActionMessage::Check, data));
                 }
